Revert change to TimeSyncMessage.Tick from int to long (for some reason it broke some tests...), revert some debug logs

+++ com.unity.netcode.gameobjects/Runtime/Messaging/Messages/TimeSyncMessage.cs
@@ -2,7 +2,7 @@ namespace Unity.Netcode
 {
     internal struct TimeSyncMessage : INetworkMessage, INetworkSerializeByMemcpy
     {
-        public long Tick;
+        public int Tick;
 
         public void Serialize(FastBufferWriter writer)
         {
